+++ src/minimax.h
@@ -1,0 +1,5 @@
+
+#ifndef TICTACTOE_MINIMAX_H
+#define TICTACTOE_MINIMAX_H
+
+#endif //TICTACTOE_MINIMAX_H
