+++ src/nextstates.h
@@ -1,0 +1,12 @@
+
+#ifndef TICTACTOE_NEXTSTATES_H
+#define TICTACTOE_NEXTSTATES_H
+
+
+struct nextstates {
+    struct state states[SIZE*SIZE];
+    int n; // number of next states
+};
+
+
+#endif //TICTACTOE_NEXTSTATES_H
