+++ src/player.c
@@ -1,0 +1,7 @@
+
+#include "player.h"
+
+enum Player other(enum Player player) {
+    if (player == Player2) return Player1;
+    return Player2;
+}
