+++ src/player.h
@@ -1,0 +1,9 @@
+
+#ifndef TICTACTOE_PLAYER_H
+#define TICTACTOE_PLAYER_H
+
+enum Player {Player1 = 'x', Player2 = 'o'};
+
+enum Player other(enum Player);
+
+#endif //TICTACTOE_PLAYER_H
