+++ src/state.h
@@ -1,0 +1,10 @@
+#include "constants.h"
+
+#ifndef TICTACTOE_STATE_H
+#define TICTACTOE_STATE_H
+
+struct state {
+    char arr[SIZE][SIZE];;
+};
+
+#endif //TICTACTOE_STATE_H
